requirements.txt
discord.py>=2.3.2
python-dateutil>=2.9.0
pytz>=2024.1
yt-dlp>=2024.7.1
PyNaCl>=1.5.0
spotipy>=2.24.0
psutil>=5.9
emoji>=2.10.0
